Fix permute
