+++ sharktank/sharktank/ops/shaping/permute.py
@@ -6,10 +6,7 @@
 
 from typing import List
 import torch
-from sharktank.types import (
-    AnyTensor,
-    unbox_tensor,
-)
+from sharktank.types import AnyTensor, unbox_tensor, SplitPrimitiveTensor
 from sharktank.ops._registry import overridable
 
 
@@ -25,3 +22,10 @@ def permute(tensor: AnyTensor, dims: List[int]) -> AnyTensor:
 def permute(tensor: torch.Tensor, dims: List[int]):
     torch_tensor = unbox_tensor(tensor)
     return torch.permute(torch_tensor, dims)
+
+
+@permute.override(SplitPrimitiveTensor)
+def permute_split(tensor: SplitPrimitiveTensor, dims: List[int]):
+    permuted_shards = [permute(shard, dims) for shard in tensor.shards]
+    permuted_shard_dim = dims[tensor.shard_dim]
+    return SplitPrimitiveTensor(ts=permuted_shards, shard_dim=permuted_shard_dim)
